reorder files, separated new subpackages
