+++ hima/experiments/temporal_pooling/new/data/__init__.py
@@ -3,11 +3,3 @@
 #  All rights reserved.
 #
 #  Licensed under the AGPLv3 license. See LICENSE in the project root for license information.
-from typing import Any
-
-
-def rename_dict_keys(d: dict[str, Any], add_prefix):
-    return {
-        f'{add_prefix}{k}': d[k]
-        for k in d
-    }
