+++ hima/experiments/temporal_pooling/new/sdr_seq_cross_stats.py
@@ -9,7 +9,7 @@ import numpy as np
 
 from hima.common.sdr import SparseSdr
 from hima.common.sds import Sds
-from hima.experiments.temporal_pooling.new.metrics import (
+from hima.experiments.temporal_pooling.new.stats.metrics import (
     standardize_sample_distribution,
     sequence_similarity_elementwise, distribution_similarity, DISTR_SIM_PMF,
     DISTR_SIM_KL
