+++ hima/experiments/temporal_pooling/new/stp/__init__.py
@@ -1,0 +1,5 @@
+#  Copyright (c) 2022 Autonomous Non-Profit Organization "Artificial Intelligence Research
+#  Institute" (AIRI); Moscow Institute of Physics and Technology (National Research University).
+#  All rights reserved.
+#
+#  Licensed under the AGPLv3 license. See LICENSE in the project root for license information.
